chore: use of ContractRequest in edrs API + old edrs schema and migration removal
