+++ edc-extensions/edr/edr-api-v2/src/test/java/org/eclipse/tractusx/edc/api/edr/v2/TestFunctions.java
@@ -21,47 +21,17 @@ package org.eclipse.tractusx.edc.api.edr.v2;
 
 import jakarta.json.Json;
 import jakarta.json.JsonObject;
-import org.eclipse.edc.connector.controlplane.api.management.contractnegotiation.model.ContractOfferDescription;
 import org.eclipse.edc.connector.controlplane.contract.spi.types.negotiation.ContractNegotiation;
-import org.eclipse.edc.connector.controlplane.contract.spi.types.offer.ContractOffer;
-import org.eclipse.edc.policy.model.Policy;
-import org.eclipse.tractusx.edc.api.edr.v2.dto.NegotiateEdrRequestDto;
-import org.eclipse.tractusx.edc.edr.spi.types.NegotiateEdrRequest;
-
-import java.util.UUID;
+import org.eclipse.edc.connector.controlplane.contract.spi.types.negotiation.ContractRequest;
 
 import static org.eclipse.edc.jsonld.spi.JsonLdKeywords.TYPE;
 import static org.eclipse.edc.spi.constants.CoreConstants.EDC_NAMESPACE;
 
 public class TestFunctions {
 
-    public static ContractOfferDescription createOffer(String offerId, String assetId) {
-        return ContractOfferDescription.Builder.newInstance()
-                .offerId(offerId)
-                .assetId(assetId)
-                .policy(Policy.Builder.newInstance().build())
-                .build();
-    }
-
-    public static ContractOfferDescription createOffer(Policy policy) {
-        return ContractOfferDescription.Builder.newInstance()
-                .offerId(UUID.randomUUID().toString())
-                .assetId(UUID.randomUUID().toString())
-                .policy(policy)
-                .build();
-    }
-
-    public static ContractOfferDescription createOffer(String offerId) {
-        return createOffer(offerId, UUID.randomUUID().toString());
-    }
-
-    public static ContractOfferDescription createOffer() {
-        return createOffer(UUID.randomUUID().toString(), UUID.randomUUID().toString());
-    }
-
     public static JsonObject negotiationRequest() {
         return Json.createObjectBuilder()
-                .add(TYPE, NegotiateEdrRequestDto.EDR_REQUEST_DTO_TYPE)
+                .add(TYPE, ContractRequest.CONTRACT_REQUEST_TYPE)
                 .add(EDC_NAMESPACE + "connectorId", "test")
                 .add(EDC_NAMESPACE + "providerId", "test")
                 .add(EDC_NAMESPACE + "connectorAddress", "test")
@@ -71,18 +41,6 @@ public class TestFunctions {
                         .add(EDC_NAMESPACE + "assetId", "assetId")
                         .add(EDC_NAMESPACE + "policy", Json.createObjectBuilder().build())
                 )
-                .build();
-    }
-
-    public static NegotiateEdrRequest openRequest() {
-        return NegotiateEdrRequest.Builder.newInstance()
-                .connectorAddress("test")
-                .connectorId("id")
-                .protocol("test-protocol")
-                .offer(ContractOffer.Builder.newInstance()
-                        .id("offerId")
-                        .assetId("assetId")
-                        .policy(Policy.Builder.newInstance().build()).build())
                 .build();
     }
