--- edc-extensions/postgresql-migration/src/main/resources/org/eclipse/tractusx/edc/postgresql/migration/edr/V0_0_4__Add_ContractNegotiationId_Column.sql
@@ -1,19 +1,0 @@
---
---  Copyright (c) 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
---
---  This program and the accompanying materials are made available under the
---  terms of the Apache License, Version 2.0 which is available at
---  https://www.apache.org/licenses/LICENSE-2.0
---
---  SPDX-License-Identifier: Apache-2.0
---
---  Contributors:
---       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
---
-
--- Statements are designed for and tested with Postgres only!
-
-
--- add column
-ALTER TABLE edc_edr_cache ADD COLUMN contract_negotiation_id VARCHAR;
-
